feat(subtitles): separated web from native and updated imports
