+++ react/features/subtitles/components/web/LanguageListItem.tsx
@@ -2,9 +2,9 @@ import React, { useCallback } from 'react';
 import { WithTranslation } from 'react-i18next';
 import { makeStyles } from 'tss-react/mui';
 
-import { translate } from '../../base/i18n/functions';
-import Icon from '../../base/icons/components/Icon';
-import { IconCheck } from '../../base/icons/svg';
+import { translate } from '../../../base/i18n/functions';
+import Icon from '../../../base/icons/components/Icon';
+import { IconCheck } from '../../../base/icons/svg';
 
 interface ILanguageListItemProps extends WithTranslation {
